+++ sfml-util/sfml-util.cpp
@@ -293,9 +293,9 @@ namespace sfml_util
             for (unsigned int x(0); x < image.getSize().x; ++x)
             {
                 sf::Color color(image.getPixel(x, y));
-                color.r = static_cast<sf::Uint8>(255) - color.r;
-                color.g = static_cast<sf::Uint8>(255) - color.g;
-                color.b = static_cast<sf::Uint8>(255) - color.b;
+                color.r = static_cast<sf::Uint8>(255 - static_cast<int>(color.r));
+                color.g = static_cast<sf::Uint8>(255 - static_cast<int>(color.g));
+                color.b = static_cast<sf::Uint8>(255 - static_cast<int>(color.b));
 
                 if (WILL_INVERT_ALPHA)
                     color.a = 255 - color.a;
